requirements.txt
streamlit
requests
textblob
plotly
python-dotenv
pandas
matplotlib.pyplot
wordcloud